requirements.txt
numpy==1.23.5
spacy==3.6
python-docx==1.0.1
pdfminer.six==20221105
supabase==2.0.3
requests==2.31.0
thinc==8.1.10
# Download spacy model
# Run: python -m spacy download en_core_web_sm
# Run: python -m spacy download en_core_web_md